+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
    parameter. `TransactionContext` provides the public key of transaction
    author, ID of current service, and transaction hash (#943)
 
+- `Transaction::verify` method has been removed. (#???)
+
 - Every transaction that contains the public key of the author was refactored
    to use the author indicated in `TransactionContext`. (#984 #980 #979 #975 #971)